Break the symbolic query description datatypes into a separate module, moving SBV specific code into `Cryptol.Symbolic.SBV`
